refactor: move to utils and add nationalId back for death
